more validations for copy project by sharing

+++ app.py
@@ -167,11 +167,18 @@ class validate_project_student_file(object):
             field.data.save(path_to_student_file_stored + student_file_filename)
             student_file_workbook = load_workbook(path_to_student_file_stored + student_file_filename)
             student_file_worksheet = student_file_workbook['Sheet1']
+            find_Student = True if 'Student' in [x.value for x in
+                                                 list(student_file_worksheet.iter_rows())[0]] else False
+            find_Email = True if 'Email' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
             find_group = True if 'group' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
             find_meta_group = True if 'meta' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
             if find_group is False:
                 # os.remove(path_to_student_file_stored)
                 raise ValidationError("Can not find group")
+            elif find_Student is False:
+                raise ValidationError("Can not find Student")
+            elif find_Email is False:
+                raise ValidationError("Can not find Email")
             # os.remove(path_to_student_file_stored+student_file_filename)
             elif find_meta_group is False:
                 raise ValidationError("Can not find meta - group")
@@ -784,7 +791,7 @@ def create_project_by_share(project_id):
     new_project_name = request.form['project_name']
     duplicate_project_name = Project.query.filter_by(project_name=new_project_name, owner=current_user.username).first()
     if duplicate_project_name is not None:
-        msg = "This rubric name has been used before"
+        return redirect('account', msg="This rubric name has been used before")
     path_to_current_user_project = "{}/{}/{}".format(base_directory, current_user.username, new_project_name)
     os.mkdir(path_to_current_user_project)
     new_project_desc = request.form['project_desc']
@@ -795,17 +802,32 @@ def create_project_by_share(project_id):
     #check if the student file is valid:
     student_file_workbook = load_workbook(path_to_student_file_stored)
     student_file_worksheet = student_file_workbook['Sheet1']
+    find_Student = True if 'Student' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
+    find_Email = True if 'Email' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
     find_group = True if 'group' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
-    if find_group is False:
-        raise ValidationError("Can not find group information")
+    find_meta_group = True if 'meta' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
 
+    if find_Student is False:
+        return redirect('account', msg="no Student column in student file!")
+    if find_Email is False:
+        return redirect('account', msg="no Email column in student file!")
+    if find_group is False:
+        return redirect('account', msg="no group column in student file!")
+    if find_meta_group is False:
+        return redirect('account', msg="no meta group column in student file!")
     # copy json file:
     project = Permission.query.filter_by(project_id=project_id).first()
-    owner = project.owner
-    project_name = project.project
-    path_to_json_file = "{}/{}/{}/TW.json".format(base_directory, owner, project_name)#jacky: use project name and project owner info to locate the path of json?
-    path_to_json_file_stored = "{}/TW.json".format(path_to_current_user_project)
-    shutil.copy2(path_to_json_file, path_to_current_user_project)
+    if project is not None:
+        owner = project.owner
+        project_name = project.project
+        path_to_json_file = "{}/{}/{}/TW.json".format(base_directory, owner, project_name)#jacky: use project name and project owner info to locate the path of json?
+        path_to_json_file_stored = "{}/TW.json".format(path_to_current_user_project)
+        if os.path.exists(path_to_json_file):
+            shutil.copy2(path_to_json_file, path_to_json_file_stored)
+        else:
+            return redirect('account', msg="the rubric you were trying to copy has been deleted")
+    else:
+        return redirect('account', msg="the rubric you were trying to copy has been deleted")
 
     #create project:
     # create group file depending on student file
@@ -904,9 +926,19 @@ def create_project_by_share_name_and_owner(type, project_name, project_owner):
     #check if the student file is valid:
     student_file_workbook = load_workbook(path_to_student_file_stored)
     student_file_worksheet = student_file_workbook['Sheet1']
+    find_Student = True if 'Student' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
+    find_Email = True if 'Email' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
     find_group = True if 'group' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
+    find_meta_group = True if 'meta' in [x.value for x in list(student_file_worksheet.iter_rows())[0]] else False
+
+    if find_Student is False:
+        return redirect('account', msg="no Student column in student file!")
+    if find_Email is False:
+        return redirect('account', msg="no Email column in student file!")
     if find_group is False:
-        raise ValidationError("Can not find group information")
+        return redirect('account', msg="no group column in student file!")
+    if find_meta_group is False:
+        return redirect('account', msg="no meta group column in student file!")
 
     # copy json file:
     # project = Permission.query.filter_by(project_id=project_id).first()
@@ -917,8 +949,10 @@ def create_project_by_share_name_and_owner(type, project_name, project_owner):
     else:
         path_to_json_file = "{}/{}/{}".format(home_directory, "Default", project_name)
     path_to_json_file_stored = "{}/TW.json".format(path_to_current_user_project)
-    shutil.copy2(path_to_json_file, path_to_json_file_stored)
-
+    if os.path.exists(path_to_json_file):
+        shutil.copy2(path_to_json_file, path_to_json_file_stored)
+    else:
+        return redirect('account', msg="the rubric you were trying to copy has been deleted")
     #create project:
     # create group file depending on student file
     list_of_group = select_by_col_name('group', student_file_worksheet)
@@ -1766,9 +1800,12 @@ def sendEmail(project_id, evaluation_name):
     return redirect(url_for('project_profile', project_id=project_id))
 
 
-@app.route('/account', methods=['GET', 'POST'])
+@app.route('/account/<string:msg>', methods=['GET', 'POST'])
 @login_required
-def account():
+def account(msg):
+    #if msg is success, msg = ""
+    if msg == "success":
+        msg = ""
     # if the notification is "permission"
     # notifications = Notification.query.filter_by(to_user=current_user.username).all()
 
@@ -1780,7 +1817,7 @@ def account():
         with open(path_to_this_json, 'r') as file:
             json_data_of_current_json_file = json.loads(file.read(), strict=False)
         json_data_of_all_default_rubric[json_file.json_name] = json_data_of_current_json_file
-    return render_template('account.html', default_json_list=json_list, json_data_of_all_default_rubric=json_data_of_all_default_rubric)
+    return render_template('account.html', msg=msg, default_json_list=json_list, json_data_of_all_default_rubric=json_data_of_all_default_rubric)
 
 @app.route('/search_project', methods=['POST'])
 @login_required
